requirements.txt
yfinance==0.2.55
pandas==2.2.3
numpy==1.26.4
flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
